+++ integration-tests/steps/change_sets.py
@@ -1,7 +1,8 @@
 from behave import *
 import time
 
-from sceptre.config.reader import ConfigReader
+from sceptre.plan.plan import SceptrePlan
+from sceptre.context import SceptreContext
 from botocore.exceptions import ClientError
 from helpers import read_template_file, get_cloudformation_stack_name
 from helpers import retry_boto_call
@@ -48,11 +49,15 @@ def step_impl(context, stack_name):
 
 @when('the user creates change set "{change_set_name}" for stack "{stack_name}"')
 def step_impl(context, change_set_name, stack_name):
-    config_reader = ConfigReader(context.sceptre_dir)
-    stack = config_reader.construct_stack(stack_name + ".yaml")
+    sceptre_context = SceptreContext(
+        command_path=stack_name + '.yaml',
+        project_path=context.sceptre_dir
+    )
+
+    sceptre_plan = SceptrePlan(sceptre_context)
     allowed_errors = {'ValidationError', 'ChangeSetNotFound'}
     try:
-        stack.create_change_set(change_set_name)
+        sceptre_plan.create_change_set(change_set_name)
     except ClientError as e:
         if e.response['Error']['Code'] in allowed_errors:
             context.error = e
@@ -64,11 +69,15 @@ def step_impl(context, change_set_name, stack_name):
 
 @when('the user deletes change set "{change_set_name}" for stack "{stack_name}"')
 def step_impl(context, change_set_name, stack_name):
-    config_reader = ConfigReader(context.sceptre_dir)
-    stack = config_reader.construct_stack(stack_name + ".yaml")
+    sceptre_context = SceptreContext(
+        command_path=stack_name + '.yaml',
+        project_path=context.sceptre_dir
+    )
+
+    sceptre_plan = SceptrePlan(sceptre_context)
     allowed_errors = {'ValidationError', 'ChangeSetNotFound'}
     try:
-        stack.delete_change_set(change_set_name)
+        sceptre_plan.delete_change_set(change_set_name)
     except ClientError as e:
         if e.response['Error']['Code'] in allowed_errors:
             context.error = e
@@ -79,27 +88,35 @@ def step_impl(context, change_set_name, stack_name):
 
 @when('the user lists change sets for stack "{stack_name}"')
 def step_impl(context, stack_name):
-    config_reader = ConfigReader(context.sceptre_dir)
-    stack = config_reader.construct_stack(stack_name + ".yaml")
+    sceptre_context = SceptreContext(
+        command_path=stack_name + '.yaml',
+        project_path=context.sceptre_dir
+    )
+
+    sceptre_plan = SceptrePlan(sceptre_context)
     allowed_errors = {'ValidationError', 'ChangeSetNotFound'}
     try:
-        response = stack.list_change_sets()
+        sceptre_plan.list_change_sets()
     except ClientError as e:
         if e.response['Error']['Code'] in allowed_errors:
             context.error = e
             return
         else:
             raise e
-    context.output = response
+    context.output = sceptre_plan.responses[0]
 
 
 @when('the user executes change set "{change_set_name}" for stack "{stack_name}"')
 def step_impl(context, change_set_name, stack_name):
-    config_reader = ConfigReader(context.sceptre_dir)
-    stack = config_reader.construct_stack(stack_name + ".yaml")
+    sceptre_context = SceptreContext(
+        command_path=stack_name + '.yaml',
+        project_path=context.sceptre_dir
+    )
+
+    sceptre_plan = SceptrePlan(sceptre_context)
     allowed_errors = {'ValidationError', 'ChangeSetNotFound'}
     try:
-        stack.execute_change_set(change_set_name)
+        sceptre_plan.execute_change_set(change_set_name)
     except ClientError as e:
         if e.response['Error']['Code'] in allowed_errors:
             context.error = e
@@ -110,18 +127,22 @@ def step_impl(context, change_set_name, stack_name):
 
 @when('the user describes change set "{change_set_name}" for stack "{stack_name}"')
 def step_impl(context, change_set_name, stack_name):
-    config_reader = ConfigReader(context.sceptre_dir)
-    stack = config_reader.construct_stack(stack_name + ".yaml")
+    sceptre_context = SceptreContext(
+        command_path=stack_name + '.yaml',
+        project_path=context.sceptre_dir
+    )
+
+    sceptre_plan = SceptrePlan(sceptre_context)
     allowed_errors = {'ValidationError', 'ChangeSetNotFound'}
     try:
-        response = stack.describe_change_set(change_set_name)
+        sceptre_plan.describe_change_set(change_set_name)
     except ClientError as e:
         if e.response['Error']['Code'] in allowed_errors:
             context.error = e
             return
         else:
             raise e
-    context.output = response
+    context.output = sceptre_plan.responses[0]
 
 
 @then('stack "{stack_name}" has change set "{change_set_name}" in "{state}" state')
