+++ sceptre/cli/create.py
@@ -22,17 +22,17 @@ def create_command(ctx, path, change_set_name, yes):
     creates a change set for stack in PATH.
     """
     context = SceptreContext(
-                command_path=path,
-                project_path=ctx.obj.get("project_path"),
-                user_variables=ctx.obj.get("user_variables"),
-                options=ctx.obj.get("options")
-            )
+        command_path=path,
+        project_path=ctx.obj.get("project_path"),
+        user_variables=ctx.obj.get("user_variables"),
+        options=ctx.obj.get("options")
+    )
 
     action = "create"
     plan = SceptrePlan(context)
 
-  # TODO this isn't going to work with sub-stack-group-stacks
-    if any(change_set_name for change_set_name in plan.stack_group.stacks):
+    # TODO this isn't going to work with sub-stack-group-stacks
+    if change_set_name:
         confirmation(action, yes, change_set=change_set_name, stack=path)
         plan.create_change_set(change_set_name)
     else:
