+++ sceptre/cli/helpers.py
@@ -73,6 +73,7 @@ def write(var, output_format="str", no_colour=True):
     :type no_colour: bool
     """
     stream = var
+
     if output_format == "json":
         encoder = CustomJsonEncoder()
         stream = encoder.encode(var)
@@ -82,6 +83,8 @@ def write(var, output_format="str", no_colour=True):
         stream = var
     if not no_colour:
         stack_status_colourer = StackStatusColourer()
+        if not isinstance(var, str):
+            stream = str(var)
         stream = stack_status_colourer.colour(stream)
 
     click.echo(stream)
