+++ sceptre/cli/list.py
@@ -2,9 +2,9 @@ import click
 
 from sceptre.context import SceptreContext
 from sceptre.cli.helpers import (
-          catch_exceptions,
-          write
-        )
+    catch_exceptions,
+    write
+)
 from sceptre.plan.plan import SceptrePlan
 
 
@@ -27,13 +27,12 @@ def list_resources(ctx, path):
 
     """
     context = SceptreContext(
-                command_path=path,
-                project_path=ctx.obj.get("project_path"),
-                user_variables=ctx.obj.get("user_variables"),
-                options=ctx.obj.get("options"),
-                output_format=ctx.obj.get("output_format")
-            )
-
+        command_path=path,
+        project_path=ctx.obj.get("project_path"),
+        user_variables=ctx.obj.get("user_variables"),
+        options=ctx.obj.get("options"),
+        output_format=ctx.obj.get("output_format")
+    )
     plan = SceptrePlan(context)
     plan.describe_resources()
     write(plan.responses, context.output_format)
@@ -53,27 +52,25 @@ def list_outputs(ctx, path, export):
 
     """
     context = SceptreContext(
-                command_path=path,
-                project_path=ctx.obj.get("project_path", None),
-                user_variables=ctx.obj.get("user_variables", {}),
-                options=ctx.obj.get("options", {}),
-                output_format=ctx.obj.get("output_format", {})
-            )
+        command_path=path,
+        project_path=ctx.obj.get("project_path", None),
+        user_variables=ctx.obj.get("user_variables", {}),
+        options=ctx.obj.get("options", {}),
+        output_format=ctx.obj.get("output_format", {})
+    )
 
     plan = SceptrePlan(context)
     plan.describe_outputs()
 
     if export == "envvar":
         write("\n".join(
-            [
-                "export SCEPTRE_{0}={1}".format(
-                    output["OutputKey"], output["OutputValue"]
-                )
-                for output in plan.responses
-            ]
+            "export SCEPTRE_{0}={1}".format(
+                output["OutputKey"], output["OutputValue"]
+            )
+            for output in plan.responses[0]
         ))
     else:
-        write(plan.responses, context.output_format)
+        write(plan.responses[0], context.output_format)
 
 
 @list_group.command(name="change-sets")
@@ -86,12 +83,12 @@ def list_change_sets(ctx, path):
 
     """
     context = SceptreContext(
-                command_path=path,
-                project_path=ctx.obj.get("project_path"),
-                user_variables=ctx.obj.get("user_variables"),
-                output_format=ctx.obj.get("output_format"),
-                options=ctx.obj.get("options")
-            )
+        command_path=path,
+        project_path=ctx.obj.get("project_path"),
+        user_variables=ctx.obj.get("user_variables"),
+        output_format=ctx.obj.get("output_format"),
+        options=ctx.obj.get("options")
+    )
 
     plan = SceptrePlan(context)
     plan.list_change_sets()
