+++ tests/test_cli.py
@@ -9,15 +9,15 @@ from mock import MagicMock, patch, sentinel
 import pytest
 import click
 
-from sceptre.context import SceptreContext
 from sceptre.cli import cli
 from sceptre.config.reader import ConfigReader
 from sceptre.stack import Stack
 from sceptre.stack_group import StackGroup
+from sceptre.plan.actions import StackActions
+from sceptre.plan.actions import StackGroupActions
 from sceptre.stack_status import StackStatus
 from sceptre.cli.helpers import setup_logging, write, ColouredFormatter
 from sceptre.cli.helpers import CustomJsonEncoder, catch_exceptions
-from sceptre.cli.helpers import get_stack_or_stack_group
 from botocore.exceptions import ClientError
 from sceptre.exceptions import SceptreException
 
@@ -25,27 +25,44 @@ from sceptre.exceptions import SceptreException
 class TestCli(object):
 
     def setup_method(self, test_method):
-        self.patcher_ConfigReader = patch("sceptre.cli.helpers.ConfigReader")
-        self.patcher_getcwd = patch("sceptre.cli.os.getcwd")
+        self.patcher_ConfigReader = patch("sceptre.plan.plan.ConfigReader")
+        self.patcher_StackActions = patch("sceptre.plan.executor.StackActions")
+        self.patcher_StackGroupActions = patch(
+            "sceptre.plan.actions.StackGroupActions")
 
         self.mock_ConfigReader = self.patcher_ConfigReader.start()
-        self.mock_getcwd = self.patcher_getcwd.start()
+        self.mock_StackActions = self.patcher_StackActions.start()
+        self.mock_StackGroupActions = self.patcher_StackGroupActions.start()
 
         self.mock_config_reader = MagicMock(spec=ConfigReader)
+        self.mock_stack_actions = MagicMock(spec=StackActions)
+        self.mock_stack_group_actions = MagicMock(spec=StackGroupActions)
+
         self.mock_stack = MagicMock(spec=Stack)
         self.mock_stack_group = MagicMock(spec=StackGroup)
+
+        self.mock_stack.name = 'mock-stack'
+        self.mock_stack.region = None
+        self.mock_stack.profile = None
+        self.mock_stack.external_name = None
+        self.mock_stack_group.stacks = [self.mock_stack]
+
         self.mock_config_reader.construct_stack.return_value = self.mock_stack
         self.mock_config_reader.construct_stack_group.return_value = \
             self.mock_stack_group
 
+        self.mock_stack_actions.stack = self.mock_stack
+        self.mock_stack_group_actions.stack_group = self.mock_stack_group
+
         self.mock_ConfigReader.return_value = self.mock_config_reader
-        self.mock_getcwd.return_value = sentinel.cwd
+        self.mock_StackActions.return_value = self.mock_stack_actions
+        self.mock_StackGroupActions.return_value = self.mock_stack_actions
 
         self.runner = CliRunner()
 
     def teardown_method(self, test_method):
         self.patcher_ConfigReader.stop()
-        self.patcher_getcwd.stop()
+        self.patcher_StackActions.stop()
 
     @patch("sys.exit")
     def test_catch_excecptions(self, mock_exit):
@@ -105,28 +122,26 @@ class TestCli(object):
         def noop(ctx):
             click.echo(yaml.safe_dump(ctx.obj.get("user_variables")))
 
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             for name, content in files.items():
                 with open(name, "w") as fh:
                     yaml.safe_dump(content, fh)
 
             result = self.runner.invoke(cli, command)
-        self.patcher_getcwd.start()
 
         user_variables = yaml.safe_load(result.output)
         assert result.exit_code == 0
         assert user_variables == output
 
     def test_validate_template_with_valid_template(self):
-        self.mock_stack.template.validate.return_value = {
-                    "Parameters": "Example",
-                    "ResponseMetadata": {
-                        "HTTPStatusCode": 200
-                    }
-                }
+        self.mock_stack_actions.validate.return_value = {
+            "Parameters": "Example",
+            "ResponseMetadata": {
+                "HTTPStatusCode": 200
+            }
+        }
         result = self.runner.invoke(cli, ["validate", "dev/vpc.yaml"])
-        self.mock_stack.template.validate.assert_called_with()
+        self.mock_stack_actions.validate.assert_called_with()
 
         assert result.output == "Template is valid. Template details:\n\n" \
             "Parameters: Example\n\n"
@@ -142,24 +157,24 @@ class TestCli(object):
             },
             "ValidateTemplate"
         )
-        self.mock_stack.template.validate.side_effect = client_error
+        self.mock_stack_actions.validate.side_effect = client_error
 
         expected_result = str(client_error) + "\n"
         result = self.runner.invoke(cli, ["validate", "dev/vpc.yaml"])
         assert result.output == expected_result
 
     def test_estimate_template_cost_with_browser(self):
-        self.mock_stack.template.estimate_cost.return_value = {
-                "Url": "http://example.com",
-                "ResponseMetadata": {
-                    "HTTPStatusCode": 200
-                }
+        self.mock_stack_actions.estimate_cost.return_value = {
+            "Url": "http://example.com",
+            "ResponseMetadata": {
+                "HTTPStatusCode": 200
             }
+        }
 
         args = ["estimate-cost", "dev/vpc.yaml"]
         result = self.runner.invoke(cli, args)
 
-        self.mock_stack.template.estimate_cost.assert_called_with()
+        self.mock_stack_actions.estimate_cost.assert_called_with()
 
         assert result.output == \
             '{0}{1}'.format("View the estimated cost at:\n",
@@ -176,13 +191,14 @@ class TestCli(object):
             },
             "Webbrowser"
         )
-        self.mock_stack.template.estimate_cost.side_effect = client_error
+        self.mock_stack_actions.estimate_cost.side_effect = client_error
 
         expected_result = str(client_error) + "\n"
         result = self.runner.invoke(
-                    cli,
-                    ["estimate-cost", "dev/vpc.yaml"]
-                )
+            cli,
+            ["estimate-cost", "dev/vpc.yaml"]
+        )
+
         assert result.output == expected_result
 
     def test_lock_stack(self):
@@ -190,18 +206,18 @@ class TestCli(object):
             cli, ["set-policy", "dev/vpc.yaml", "-b", "deny-all"]
         )
         self.mock_config_reader.construct_stack.assert_called_with(
-                "dev/vpc.yaml"
+            "dev/vpc.yaml"
         )
-        self.mock_stack.lock.assert_called_with()
+        self.mock_stack_actions.lock.assert_called_with()
 
     def test_unlock_stack(self):
         self.runner.invoke(
             cli, ["set-policy", "dev/vpc.yaml", "-b", "allow-all"]
         )
         self.mock_config_reader.construct_stack.assert_called_with(
-                "dev/vpc.yaml"
+            "dev/vpc.yaml"
         )
-        self.mock_stack.unlock.assert_called_with()
+        self.mock_stack_actions.unlock.assert_called_with()
 
     def test_set_policy_with_file_flag(self):
         policy_file = "tests/fixtures/stack_policies/lock.json"
@@ -211,7 +227,7 @@ class TestCli(object):
         assert result.exit_code == 0
 
     def test_describe_policy_with_existing_policy(self):
-        self.mock_stack.get_policy.return_value = {
+        self.mock_stack_actions.get_policy.return_value = {
             "StackPolicyBody": "Body"
         }
 
@@ -240,23 +256,24 @@ class TestCli(object):
                 ]
             }
         }
-        self.mock_stack_group.describe_resources.return_value = response
+        self.mock_stack_actions.describe_resources.return_value = response
         result = self.runner.invoke(cli, ["list", "resources", "dev"])
-        assert yaml.load(result.output) == response
+
+        assert yaml.load(result.output) == [response]
         assert result.exit_code == 0
 
     def test_list_stack_resources(self):
         response = {
-                "StackResources": [
-                    {
-                        "LogicalResourceId": "logical-resource-id",
-                        "PhysicalResourceId": "physical-resource-id"
-                    }
-                ]
-            }
-        self.mock_stack.describe_resources.return_value = response
+            "StackResources": [
+                {
+                    "LogicalResourceId": "logical-resource-id",
+                    "PhysicalResourceId": "physical-resource-id"
+                }
+            ]
+        }
+        self.mock_stack_actions.describe_resources.return_value = response
         result = self.runner.invoke(cli, ["list", "resources", "dev/vpc.yaml"])
-        assert yaml.load(result.output) == response
+        assert yaml.load(result.output) == [response]
         assert result.exit_code == 0
 
     @pytest.mark.parametrize(
@@ -280,7 +297,8 @@ class TestCli(object):
         ]
     )
     def test_stack_commands(self, command, success, yes_flag, exit_code):
-        getattr(self.mock_stack, command).return_value = \
+        run_command = getattr(self.mock_stack_actions, command)
+        run_command.return_value = \
             StackStatus.COMPLETE if success else StackStatus.FAILED
 
         kwargs = {"args": [command, "dev/vpc.yaml"]}
@@ -291,7 +309,7 @@ class TestCli(object):
 
         result = self.runner.invoke(cli, **kwargs)
 
-        getattr(self.mock_stack, command).assert_called_with()
+        run_command.assert_called_with()
         assert result.exit_code == exit_code
 
     @pytest.mark.parametrize(
@@ -315,7 +333,8 @@ class TestCli(object):
 
         result = self.runner.invoke(cli, **kwargs)
 
-        getattr(self.mock_stack, stack_command).assert_called_with("cs1")
+        getattr(self.mock_stack_actions,
+                stack_command).assert_called_with("cs1")
         assert result.exit_code == 0
 
     @pytest.mark.parametrize(
@@ -332,9 +351,7 @@ class TestCli(object):
     )
     def test_stack_group_commands(self, command, success, yes_flag, exit_code):
         status = StackStatus.COMPLETE if success else StackStatus.FAILED
-        response = {"stack": status}
-
-        getattr(self.mock_stack_group, command).return_value = response
+        getattr(self.mock_stack_actions, command).return_value = status
 
         kwargs = {"args": [command, "dev"]}
         if yes_flag:
@@ -344,7 +361,7 @@ class TestCli(object):
 
         result = self.runner.invoke(cli, **kwargs)
 
-        getattr(self.mock_stack_group, command).assert_called_with()
+        getattr(self.mock_stack_actions, command).assert_called_with()
         assert result.exit_code == exit_code
 
     @pytest.mark.parametrize(
@@ -380,7 +397,7 @@ class TestCli(object):
         if verbose_flag:
             args.append("-v")
 
-        self.mock_stack.describe_change_set.return_value = response
+        self.mock_stack_actions.describe_change_set.return_value = response
         result = self.runner.invoke(cli, args)
         if not verbose_flag:
             del response["VerboseProperty"]
@@ -389,7 +406,7 @@ class TestCli(object):
         assert result.exit_code == 0
 
     def test_list_change_sets_with_200(self):
-        self.mock_stack.list_change_sets.return_value = {
+        self.mock_stack_actions.list_change_sets.return_value = {
             "ResponseMetadata": {
                 "HTTPStatusCode": 200
             },
@@ -408,7 +425,7 @@ class TestCli(object):
             },
             "ChangeSets": "Test"
         }
-        self.mock_stack.list_change_sets.return_value = response
+        self.mock_stack_actions.list_change_sets.return_value = response
 
         result = self.runner.invoke(
             cli, ["list", "change-sets", "dev/vpc.yaml"]
@@ -417,8 +434,8 @@ class TestCli(object):
         assert yaml.load(result.output) == response
 
     def test_list_outputs(self):
-        outputs = [{"OutputKey": "Key", "OutputValue": "Value"}]
-        self.mock_stack.describe_outputs.return_value = outputs
+        outputs = {"OutputKey": "Key", "OutputValue": "Value"}
+        self.mock_stack_actions.describe_outputs.return_value = outputs
         result = self.runner.invoke(
             cli, ["list", "outputs", "dev/vpc.yaml"]
         )
@@ -427,7 +444,7 @@ class TestCli(object):
 
     def test_list_outputs_with_export(self):
         outputs = [{"OutputKey": "Key", "OutputValue": "Value"}]
-        self.mock_stack.describe_outputs.return_value = outputs
+        self.mock_stack_actions.describe_outputs.return_value = outputs
         result = self.runner.invoke(
             cli, ["list", "outputs", "dev/vpc.yaml", "-e", "envvar"]
         )
@@ -435,20 +452,21 @@ class TestCli(object):
         assert yaml.load(result.output) == "export SCEPTRE_Key=Value"
 
     def test_status_with_group(self):
-        self.mock_stack_group.describe.return_value = {"stack": "status"}
+        self.mock_stack_actions.get_status.return_value = {
+            "stack": "status"
+        }
 
         result = self.runner.invoke(cli, ["status", "dev"])
         assert result.exit_code == 0
-        assert result.output == "stack: status\n\n"
+        assert result.output == "{'stack': 'status'}\n"
 
     def test_status_with_stack(self):
-        self.mock_stack.get_status.return_value = "status"
+        self.mock_stack_actions.get_status.return_value = "status"
         result = self.runner.invoke(cli, ["status", "dev/vpc.yaml"])
         assert result.exit_code == 0
         assert result.output == "status\n"
 
     def test_init_project_non_existant(self):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             project_path = os.path.abspath('./example')
             config_dir = os.path.join(project_path, "config")
@@ -469,10 +487,8 @@ class TestCli(object):
                 config = yaml.load(config_file)
 
             assert config == defaults
-        self.patcher_getcwd.start()
 
     def test_init_project_already_exist(self):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             project_path = os.path.abspath('./example')
             config_dir = os.path.join(project_path, "config")
@@ -496,10 +512,8 @@ class TestCli(object):
             with open(os.path.join(config_dir, "config.yaml")) as config_file:
                 config = yaml.load(config_file)
             assert existing_config == config
-        self.patcher_getcwd.start()
 
     def test_init_project_another_exception(self):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             patcher_mkdir = patch("sceptre.cli.init.os.mkdir")
             mock_mkdir = patcher_mkdir.start()
@@ -507,47 +521,45 @@ class TestCli(object):
             result = self.runner.invoke(cli, ["init", "project", "example"])
             mock_mkdir = patcher_mkdir.stop()
             assert str(result.exception) == str(OSError(errno.EINVAL))
-        self.patcher_getcwd.start()
 
     @pytest.mark.parametrize(
-      "stack_group,config_structure,stdin,result", [
-        (
-         "A",
-         {"": {}},
-         'y\nA\nA\n', {"project_code": "A", "region": "A"}
-        ),
-        (
-         "A",
-         {"": {"project_code": "top", "region": "top"}},
-         'y\n\n\n', {}
-        ),
-        (
-         "A",
-         {"": {"project_code": "top", "region": "top"}},
-         'y\nA\nA\n', {"project_code": "A", "region": "A"}
-        ),
-        (
-         "A/A",
-         {
-            "": {"project_code": "top", "region": "top"},
-            "A": {"project_code": "A", "region": "A"},
-         },
-         'y\nA/A\nA/A\n', {"project_code": "A/A", "region": "A/A"}
-        ),
-        (
-         "A/A",
-         {
-            "": {"project_code": "top", "region": "top"},
-            "A": {"project_code": "A", "region": "A"},
-         },
-         'y\nA\nA\n', {}
-        )
-      ]
+        "stack_group,config_structure,stdin,result", [
+            (
+                "A",
+                {"": {}},
+                'y\nA\nA\n', {"project_code": "A", "region": "A"}
+            ),
+            (
+                "A",
+                {"": {"project_code": "top", "region": "top"}},
+                'y\n\n\n', {}
+            ),
+            (
+                "A",
+                {"": {"project_code": "top", "region": "top"}},
+                'y\nA\nA\n', {"project_code": "A", "region": "A"}
+            ),
+            (
+                "A/A",
+                {
+                    "": {"project_code": "top", "region": "top"},
+                    "A": {"project_code": "A", "region": "A"},
+                },
+                'y\nA/A\nA/A\n', {"project_code": "A/A", "region": "A/A"}
+            ),
+            (
+                "A/A",
+                {
+                    "": {"project_code": "top", "region": "top"},
+                    "A": {"project_code": "A", "region": "A"},
+                },
+                'y\nA\nA\n', {}
+            )
+        ]
     )
     def test_init_stack_group(
         self, stack_group, config_structure, stdin, result
     ):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             project_path = os.path.abspath('./example')
             config_dir = os.path.join(project_path, "config")
@@ -579,17 +591,15 @@ class TestCli(object):
 
             if result:
                 with open(os.path.join(stack_group_dir, "config.yaml"))\
-                  as config_file:
+                        as config_file:
                     config = yaml.load(config_file)
                 assert config == result
             else:
                 assert cmd_result.output.endswith(
                     "No config.yaml file needed - covered by parent config.\n"
                 )
-        self.patcher_getcwd.start()
 
     def test_init_stack_group_with_existing_folder(self):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             project_path = os.path.abspath('./example')
             config_dir = os.path.join(project_path, "config")
@@ -607,14 +617,11 @@ class TestCli(object):
                 "Do you want initialise config.yaml?"
             )
             with open(os.path.join(
-                  stack_group_dir, "config.yaml")) as config_file:
+                    stack_group_dir, "config.yaml")) as config_file:
                 config = yaml.load(config_file)
             assert config == {"project_code": "", "region": ""}
 
-        self.patcher_getcwd.start()
-
     def test_init_stack_group_with_another_exception(self):
-        self.patcher_getcwd.stop()
         with self.runner.isolated_filesystem():
             project_path = os.path.abspath('./example')
             config_dir = os.path.join(project_path, "config")
@@ -628,8 +635,6 @@ class TestCli(object):
             result = self.runner.invoke(cli, ["init", "grp", "A"])
             mock_mkdir = patcher_mkdir.stop()
             assert str(result.exception) == str(OSError(errno.EINVAL))
-
-        self.patcher_getcwd.start()
 
     def test_setup_logging_with_debug(self):
         logger = setup_logging(True, False)
@@ -688,42 +693,3 @@ class TestCli(object):
         encoder = CustomJsonEncoder()
         response = encoder.encode(datetime.datetime(2016, 5, 3))
         assert response == '"2016-05-03 00:00:00"'
-
-    def test_get_stack_or_stack_group_with_stack(self):
-        context = MagicMock(spec=SceptreContext)
-        context.project_path = "tests/fixtures"
-        context.command_path = "account/stack-group/region/vpc.yaml"
-        context.user_variables = sentinel.user_variables
-
-        stack, stack_group = get_stack_or_stack_group(context)
-        self.mock_ConfigReader.assert_called_once_with(
-            context.project_path, context.user_variables
-        )
-        assert isinstance(stack, Stack)
-        assert stack_group is None
-
-    def test_get_stack_or_stack_group_with_nested_stack(self):
-        context = MagicMock(spec=SceptreContext)
-        context.project_path = "tests/fixtures"
-        context.command_path = "account/stack-group/region/vpc.yaml"
-        context.user_variables = sentinel.user_variables
-        stack, stack_group = get_stack_or_stack_group(context)
-        self.mock_ConfigReader.assert_called_once_with(
-            context.project_path, context.user_variables
-        )
-        assert isinstance(stack, Stack)
-        assert stack_group is None
-
-    def test_get_stack_or_stack_group_with_group(self):
-        context = MagicMock(spec=SceptreContext)
-        context.project_path = "tests/fixtures"
-        context.command_path = "account/stack-group/region"
-        context.user_variables = sentinel.user_variables
-
-        stack, stack_group = get_stack_or_stack_group(context)
-
-        self.mock_ConfigReader.assert_called_once_with(
-           context.project_path, context.user_variables
-        )
-        assert isinstance(stack_group, StackGroup)
-        assert stack is None
